+++ packages/neuron-ui/src/components/ImportHardware/index.tsx
@@ -1,10 +1,10 @@
 import React, { useRef, useEffect } from 'react'
 import { Switch, Route, RouteComponentProps } from 'react-router-dom'
-import Comfirm from './confirming'
-import ImportError from './error'
+import Comfirming from './confirming'
+import ImportError from './import-error'
 import SelectModel from './select-model'
 import DetectDevice from './detect-device'
-import Success from './success'
+import Success from './import-success'
 import NameWallet from './name-wallet'
 import { RoutePath } from './common'
 
@@ -22,7 +22,7 @@ const ImportHardware = ({ match }: RouteComponentProps) => {
       <Switch>
         <Route component={SelectModel} exact path={match.url} />
         <Route component={DetectDevice} exact path={match.url + RoutePath.DetectDevice} />
-        <Route component={Comfirm} exact path={match.url + RoutePath.Comfirming} />
+        <Route component={Comfirming} exact path={match.url + RoutePath.Comfirming} />
         <Route component={ImportError} exact path={match.url + RoutePath.Error} />
         <Route component={Success} exact path={match.url + RoutePath.Success} />
         <Route component={NameWallet} exact path={match.url + RoutePath.NameWallet} />
