+++ packages/neuron-wallet/src/services/hardware/index.ts
@@ -1,138 +1,54 @@
-import type Transaction from 'models/chain/transaction'
-import WitnessArgs from 'models/chain/witness-args'
-import { serializeWitnessArgs } from '@nervosnetwork/ckb-sdk-utils'
-import Address, { AddressType } from 'models/keys/address'
-import AddressService from 'services/addresses'
-import TransactionSender from 'services/transaction-sender'
-import MultiSign from 'models/multi-sign'
-import WalletService from 'services/wallets'
+import { UnsupportedManufacturer } from 'exceptions'
+import { Hardware, DeviceInfo, Manufacturer, HardwareClass } from './common'
+import Ledger from './ledger'
 
-export abstract class Hardware {
-  public deviceInfo: DeviceInfo
-  public isConnected: boolean
-  protected firstReceiveAddress = Address.pathForReceiving(0)
+export default class HardwareWalletService {
+  private static instance: HardwareWalletService
+  private device?: Hardware
+  private supportedHardwares: Map<Manufacturer, HardwareClass> = new Map()
 
-  constructor(device: DeviceInfo) {
-    this.deviceInfo = device
-    this.isConnected = false
+  constructor () {
+    this.supportedHardwares.set(Manufacturer.Ledger, Ledger)
   }
 
-  // @TODO: After multi-signature feature is complete, refactor this function into `TransactionSender#sign`.
-  public async signTx (walletID: string, tx: Transaction, txHash: string, skipLastInputs: number = 0) {
-    const wallet = WalletService.getInstance().get(walletID)
-    const addressInfos = await AddressService.getAddressesByWalletId(walletID)
-    const witnessSigningEntries = tx.inputs.slice(0, tx.inputs.length - skipLastInputs).map((input, index) => {
-      const lockArgs: string = input.lock!.args!
-      const wit: WitnessArgs | string = tx.witnesses[index]
-      const witnessArgs: WitnessArgs = (wit instanceof WitnessArgs) ? wit : WitnessArgs.generateEmpty()
-      return {
-        witnessArgs,
-        lockHash: input.lockHash!,
-        witness: '',
-        lockArgs,
-      }
-    })
-
-    const isMultiSign = tx.inputs.length === 1 &&
-      tx.inputs[0].lock!.args.length === TransactionSender.MULTI_SIGN_ARGS_LENGTH
-
-    const multiSignBlake160s = isMultiSign ? addressInfos.map(i => {
-      return {
-        multiSignBlake160: new MultiSign().hash(i.blake160),
-        path: i.path
-      }
-    }) : []
-
-    const findPath = (args: string) => {
-      if (args.length === TransactionSender.MULTI_SIGN_ARGS_LENGTH) {
-        return multiSignBlake160s.find(i => args.slice(0, 42) === i.multiSignBlake160)!.path
-      }
-      return addressInfos.find(i => i.blake160 === args)!.path
+  public static getInstance = () => {
+    if (HardwareWalletService.instance === undefined) {
+      HardwareWalletService.instance = new HardwareWalletService()
     }
 
-    const lockHashes = new Set(witnessSigningEntries.map(w => w.lockHash))
-
-    for (const [index, lockHash] of [...lockHashes].entries()) {
-      const witnessesArgs = witnessSigningEntries.filter(w => w.lockHash === lockHash)
-      // A 65-byte empty signature used as placeholder
-      witnessesArgs[0].witnessArgs.setEmptyLock()
-
-      const path = findPath(witnessesArgs[0].lockArgs)
-
-      if (isMultiSign) {
-        const serializedWitnesses = witnessesArgs.map(value => {
-          const args = value.witnessArgs
-          if (index === 0) {
-            return args
-          }
-          if (args.lock === undefined && args.inputType === undefined && args.outputType === undefined) {
-            return '0x'
-          }
-          return serializeWitnessArgs(args.toSDK())
-        })
-        const blake160 = addressInfos.find(i => witnessesArgs[0].lockArgs.slice(0, 42) === new MultiSign().hash(i.blake160))!.blake160
-        const serializedMultiSign: string = new MultiSign().serialize(blake160)
-        const witnesses = await TransactionSender.signSingleMultiSignScript(path, serializedWitnesses, txHash, serializedMultiSign, wallet)
-        const wit = witnesses[0] as WitnessArgs
-        wit.lock = serializedMultiSign + wit.lock!.slice(2)
-        witnesses[0] = serializeWitnessArgs(wit.toSDK())
-        witnessesArgs[index].witness = witnesses[0]
-        for (let i = 0; i < witnessesArgs.length; ++i) {
-          witnessesArgs[i].witness = witnesses[i] as string
-        }
-      } else {
-        const serializedWitnesses = witnessesArgs.map(value => {
-          const args = value.witnessArgs
-          if (args.lock === undefined && args.inputType === undefined && args.outputType === undefined) {
-            return '0x'
-          }
-          return serializeWitnessArgs(args.toSDK())
-        })
-        const signture = await this.signTransaction(walletID, tx, serializedWitnesses, path)
-        witnessesArgs[index].witness = serializeWitnessArgs({
-          lock: '0x' + signture,
-          inputType: '',
-          outputType: ''
-        })
-      }
-    }
-
-    tx.witnesses = witnessSigningEntries.map(w => w.witness || '0x')
-    tx.hash = txHash
-
-    return tx
+    return HardwareWalletService.instance
   }
 
-  public abstract getExtendedPublicKey(): Promise<ExtendedPublicKey>
-  public abstract connect(hardwareInfo?: DeviceInfo): Promise<void>
-  public abstract signMessage(path: string, messageHex: string): Promise<string>
-  public abstract disconnect(): Promise<void>
-  public abstract signTransaction(walletID: string, tx: Transaction, witnesses: string[], path: string): Promise<string>
-  public abstract getAppVersion(): Promise<string>
-  public abstract getFirmwareVersion?(): Promise<string>
-}
+  public getCurrent () {
+    return this.device
+  }
 
-export type HardwareClass = new (device: DeviceInfo) => Hardware
+  public async initHardware (deviceInfo: DeviceInfo) {
+    const Device = this.supportedHardwares.get(deviceInfo.manufacturer)
+    if (!Device) {
+      throw new UnsupportedManufacturer(deviceInfo.manufacturer)
+    }
 
-export enum Manufacturer {
-  Ledger = 'Ledger'
-}
+    if (this.device?.deviceInfo.descriptor === deviceInfo.descriptor) {
+      await this.device.disconnect()
+      return this.device!
+    }
 
-export interface DeviceInfo {
-  descriptor: string
-  vendorId: string
-  manufacturer: Manufacturer
-  product: string
-  isBluetooth: boolean
-  // for single address
-  addressType: AddressType
-  addressIndex: number
-  // The following information may or may not be available to us
-  appVersion?: string
-  firmwareVersion?: string
-}
+    this.device = new Device(deviceInfo)
+    return this.device
+  }
 
-export interface ExtendedPublicKey {
-  publicKey: string
-  chainCode: string
+  public static async findDevices (device?: Pick<DeviceInfo, 'manufacturer' | 'product'>): Promise<DeviceInfo[]> {
+    const devices = await Promise.all([
+      Ledger.findDevices(),
+      // add new brand `findDevices()` here
+    ])
+    const result = devices.flat().filter(Boolean)
+
+    if (device) {
+      return result.filter(r => r.manufacturer === device.manufacturer && r.product === device.product)
+    }
+
+    return result
+  }
 }
